+++ test/general/matrix_test.cc
@@ -7,7 +7,7 @@
 using lightest::Update;
 
 namespace {
-const size_t kTestOrder{3};
+const int kTestOrder{3};
 typedef block_game::Matrix<kTestOrder> TestMatrix;
 
 bool TestDefaultConstructor() {
@@ -19,8 +19,8 @@ bool TestDefaultConstructor() {
   };
 
   bool success{true};
-  for (size_t i{0}; i < kTestOrder; ++i) {
-    for (size_t j{0}; j < kTestOrder; ++j) {
+  for (int i{0}; i < kTestOrder; ++i) {
+    for (int j{0}; j < kTestOrder; ++j) {
       if (matrix[i][j] != expected[i][j]) {
         std::cerr << "default constructor failed: ";
         std::cerr << "element [" << i << "][" << j << "] is " << matrix[i][j] << " instead of " << expected[i][j] << std::endl;
@@ -44,8 +44,8 @@ bool TestInitializerListConstructor() {
   };
 
   bool success{true};
-  for (size_t i{0}; i < kTestOrder; ++i) {
-    for (size_t j{0}; j < kTestOrder; ++j) {
+  for (int i{0}; i < kTestOrder; ++i) {
+    for (int j{0}; j < kTestOrder; ++j) {
       if (matrix[i][j] != expected[i][j]) {
         std::cerr << "initializer list constructor failed: ";
         std::cerr << "element [" << i << "][" << j << "] is " << matrix[i][j] << " instead of " << expected[i][j] << std::endl;
